requirements.txt
django~=5.0.6
psycopg2-binary
python-decouple
qrcode[pil]
django-tailwind[reload]
psycopg2-binary
pytest
pytest-django
python-decouple
djangorestframework
djangorestframework-jwt
django-htmx-autocomplete